requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.8
pyasn1==0.6.0
pycryptodome==3.20.0
pytz==2024.1
requests==2.32.3
rsa==4.9
setuptools==74.1.2
urllib3==1.26.20
